Added rake tasks...
added reload rake task to reload user counter_cache in team.
added generation of certificates as a rake task (this task generates but doesn't send the certificates, there is already a rake task that generates and sends the certificates)
upped line limit to 150 on rubocop since Team class was over the limit.