Fix issue with certificates being generated too early
This code is less efficient than the previous code however it is the best we can do in Heroku. Since Heroku workers are restarted every 24 hours we are having an issue where we have so many teams that the emails can't finish sending before the worker is reset, losing all of the messages. This modifies the code to generate the email on the fly and send it right away, avoiding the issue of the file not being available.