Improved Navbar #464
Fixes Issue #464 with Improvments: If the current user is an admin they see all the links in the navbar. If it is before a competition all navbar links should display for the admin; however, if the current user is not an admin only the Messages tab will be displayed before the competition.